bridge_bidding_advisor/requirements.txt
chromadb
python-dotenv
dspy-ai
BeautifulSoup4
langchain
